Give PIRUS::replay access to request URL
In addition:

Use direct method of retrieving Access from database.

Do not retry ping if there is not enough information to form a ping.

Actually reschedule for 24 hours' time: a return code of 0 is treated
by EPrints::DataObj::EventQueue::execute the same as HTTP_NOT_FOUND,
which means changes to the event are not committed, and if 'cleanup'
is True (default), the event is immediately deleted. HTTP_RESET_CONTENT
is the correct return value to do what is promised in the $fail_message.